+++ CHANGELOG.md
@@ -1,6 +1,16 @@
+-- 2024.07.24 - v1.0.3
+
+- feat: add compile and package creator bash script for quick deployment
+- fix: MaxLevel config was not being used
+- fix: InitialSkillpoints config was not being used
+- fix: level-settings config section not being generated
+- fix: unable to unload plugin when database is not configured
+- fix: new players not being added to the database
+- chore: change README.md to use the team's structure
+
 -- 2024.05.18 - v1.0.2
 
-fix: Threading problems on hotReload
+- fix: Threading problems on hotReload
 
 -- 2024.05.15 - v1.0.1